% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_domain_type.R
\name{adam_domain_type}
\alias{adam_domain_type}
\title{Identify the data set type of ads files by file name}
\usage{
adam_domain_type(path = NULL, keep = NULL, drop = NULL)
}
\arguments{
\item{path}{ads path}

\item{keep}{only keep the domains provided, e.g. \code{keep = 'adsl'}}

\item{drop}{exclude the domains provided, e.g. \code{drop = 'adxb'}}
}
\value{
A tibble with one row for each \code{.sas7bdat} file in the specified folder and the following columns

\item{file}{File path}
\item{type}{File type (\emph{adsl}, \emph{bds}, \emph{occds} or \emph{none})}
\item{dom}{Name of the adam domain, i.e. the file name without its extension}
}
\description{
\code{adam_domain_type()} returns the look up table that is used
for determining the type of an ads data set (ads, bds or occds).
}
\details{
Files are read from the given \code{path} and file names are matched to their corresponsing type (ads, bds or occds) using a look up table.
This information is e.g. used to determine which versions of \code{adam_spec_*} and \code{build_*} to use for further processing
Parameters \code{keep} and \code{drop} allow control over which files to use and ignore, resp.
(If both are provided files are kept if they are \code{kept} but not in \code{drop}.)

Note that this table contains regular expressions when searching for a particular domain
(\code{adqs.*} will match e.g. adqskccq and adqsnyha).
}
\section{Authors}{


Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

