% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec.R
\name{adam_spec}
\alias{adam_spec}
\title{Creating a specification for building a wide format data set from ADaM data}
\usage{
adam_spec(
  path,
  filter = NULL,
  keep = NULL,
  drop = NULL,
  pre_study = FALSE,
  attach_data = FALSE,
  id = "SUBJID",
  trt = "TRT01A",
  add_bds = NULL,
  file_ext = c("rds", "sas7bdat"),
  fct_levels = NULL,
  catalog_file = NULL
)
}
\arguments{
\item{path}{path to a directory containing ads files in .sas7bdat format}

\item{filter}{a character vector of conditions to be passed to \code{dplyr::filter()},
e.g. regarding visits, treatment arms or parameters. Defaults to NULL.}

\item{keep, drop}{character vectors controlling the subset of data sets in the given \code{path}
to create the specification for (e.g. \code{c('adsl', 'advs'))}).
If both \code{keep} and \code{drop} are specified, only \code{keep} will be used.
Both default to NULL, which means that all (known) domains are included.}

\item{pre_study}{boolean. Include only pre-study events from occurrence data sets
(see \code{\link{adam_spec_occds}()} for details). Defaults to FALSE.}

\item{attach_data}{boolean indicating whether the imported raw data is included in
the output. Defaults to FALSE.}

\item{id, trt}{id and treatment column names (see e.g. \code{\link{adam_spec_adsl}()}
for details).}

\item{add_bds}{character vector of domain names of type bds that are not included
in the package library of ADaM types (yet), but should be processed as per usual, e.g. 'adfapr'}

\item{file_ext}{only rds and sas7bdat data sets are allowed (e.g. \code{file_ext = 'rds'}). User may select
only sas7bdat, only rds or set a priorization rule (\code{file_ext = c('rds', 'sas7bdat')}, see Details).
Defaults to c('rds', 'sas7bdat'), i.e. rds if available, sas7bdat else.}

\item{fct_levels}{optional list of named vectors providing code-decode pairs
and/or setting the level order for factors in an adsl data set
(see details section of \code{\link[=adam_spec_adsl]{adam_spec_adsl()}} for structure).}

\item{catalog_file}{path to the catalog file to be passed to
\code{\link[haven:read_sas]{haven::read_sas()}} for adsl. Defaults to \code{NULL}.
Ignored if \code{file} is not a sas7bdat file.}
}
\value{
\code{adam_spec()} returns named list of specifications that can be passed to the \code{\link{build}()} function.
Each element contains the specification for a single data set and is named with the domain abbreviation (e.g. adsl, adqskccq).
The list can be manually adjusted if required, e.g. adding further specifications or altering existing ones. See the documentation
of the \verb{adam_spec_*()} for a detailed description of the output object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

\code{\link{adam_spec}()} is a wrapper for the \verb{adam_spec_*()} functions.
It creates a list of specifications on how to extract and process data from
ADaM data sets in a given location.
The resulting list can be passed to \code{\link{build}()}, where the
created specs are applied and the generated data sets are combined into a single wide format data set.
}
\details{
\code{adam_spec()} matches file names in the given path against an internal library
to decide on which \verb{adam_spec_*()} function to use for which data set.
Only files in the library will be processed, the rest will be ignored. Names of unprocessed files will be printed to the console.
For those, specifications may be created manually using the appropriate \verb{adam_spec_*()} function and appended to the specification list created by \code{adam_*_spec()}.

By specifying e.g. \code{file_ext = 'rds'}, only rds data will be considered
for building the specification. To use only sas7bdat, analogously specify
by file extension \code{file_ext = 'sas7bdat'}.
Preferred file types can be specified using a character vector
\code{file_ext = c('rds', 'sas7bdat')}: If the same file name is found in
\code{path} with both extensions, the file with the former extension is used,
the one with the latter ignored. For unambiguous file names (either only sas7bdat or only rds)
both are used.

Individual filters are only applied if the resulting data set has a positive number of rows (ignoring those causing errors or yielding a 0-row data set).

Please refer to the documentations of the \verb{adam_spec_*()} functions for full details.
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

\seealso{
\code{\link{adam_spec_adsl}()}, \code{\link{adam_spec_bds}()},  \code{\link{adam_spec_occds}()}
}
