% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_adsl.R
\name{adam_spec_adsl}
\alias{adam_spec_adsl}
\title{spec adsl}
\usage{
adam_spec_adsl(
  file,
  id = "SUBJID",
  trt = NULL,
  keep = NULL,
  drop = NULL,
  filter = NULL,
  attach_data = FALSE,
  ...
)
}
\arguments{
\item{file}{the file}

\item{id}{name of id column to keep}

\item{trt}{name of trt column to keep}

\item{keep}{columns to be kept (overrides blacklist)}

\item{drop}{superseded. columns to be dropped (overrides whitelist)}

\item{filter}{character vector of filter criteria to be evaluated}

\item{attach_data}{boolean. attach the imported raw data

adsl.sas7bdat' %in% list.files(path)}
}
\description{
spec adsl
}
