% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_bds.R
\name{adam_spec_bds}
\alias{adam_spec_bds}
\title{Create spec object for bds data sets (basic data structure)}
\usage{
adam_spec_bds(
  file,
  id = "SUBJID",
  param = NULL,
  label = NULL,
  unit = NULL,
  time = NULL,
  value = NULL,
  filter = NULL,
  attach_data = FALSE
)
}
\arguments{
\item{file}{the sas file}

\item{id}{name of id column to be kept and used for merge of data sets}

\item{param}{name of the column that identifies the parameter. Defaults to NULL, will be guessed if not set (see Details).}

\item{label}{name of the column that gives column labels. Defaults to NULL.}

\item{unit}{Defaults to NULL, will be guessed if not set (see Details).}

\item{time}{Defaults to NULL, will be guessed if not set (see Details).}

\item{value}{Defaults to NULL, will be guessed if not set (see Details).}

\item{filter}{character vector of filters to be applied to the bds data set.
Individual filters will only be considered if the resulting data set has positive number of rows. Defaults to NULL.}

\item{attach_data}{boolean. attach the imported raw data}
}
\description{
Values for arguments param, label, unit, time, value will be guessed if not provided.
Guess will be the first of the following options that matches a column name (exact match).
\itemize{
\item\link{param} 'PARAMCD', paste0(dom,'TESTCD')
\item\link{label} substring of param with the last two characters removed
\item\link{time}  'AVISIT', 'VISIT'
\item\link{value} 'AVAL',   paste0(dom, c("STRESN", "ORRES"))
\item\link{unit}  'AVALU',  paste0(dom, c("STRESU", "ORRESU"))
}
Function will escape if one of param or value are neither provided nor can be guessed.
A parameter dictionary will be created: A tibble with unique combinations of param, label, unit (or the provided subset)
}
\section{Authors}{


Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

