% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_occds.R
\name{adam_spec_occds}
\alias{adam_spec_occds}
\title{spec occ}
\usage{
adam_spec_occds(
  file,
  id = "SUBJID",
  label = NULL,
  time = NULL,
  value = NULL,
  filter = NULL,
  count = FALSE,
  pre_study = FALSE,
  attach_data = FALSE
)
}
\arguments{
\item{file}{the sas file}

\item{id}{name of id column to be kept and used for merge of data sets}

\item{label}{name of the column that identifies the occurrence labels. Defaults to NULL, will be guessed if not set (see Details).}

\item{time}{name of the column that is used for time filtering. Defaults to NULL, will be guessed if not set (see Details).}

\item{value}{optional value column (e.g. AE severity). Defaults to NULL, which leads to an Y/N coding of the event}

\item{filter}{character vector of filters to be applied to the bds data set. 
Individual filters will only be considered if the resulting data set has positive number of rows. Defaults to NULL.}

\item{attach_data}{boolean. attach the imported raw data}
}
\description{
For file names 'adae.sas7bdat', 'adcm.sas7bdat' and 'admh.sas7bdat', values for
arguments \code{label} and \code{time} will be guessed if not provided. 
Guesses will be the first of the following options that matches a column name (exact match).
}
