% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_adsl.R
\name{adjust_adsl}
\alias{adjust_adsl}
\title{Adjust spec object for data sets of type 'adsl'}
\usage{
adjust_adsl(spec, add = NULL, drop = NULL, id = "adsl")
}
\arguments{
\item{spec}{spec object to modify}

\item{add}{character vector of columns to added (in comparison to automated selection)}

\item{drop}{character vector of column names to be dropped}

\item{id}{name of list element to modify in the spec}
}
\value{
A modified version of \code{spec} to be used as input to \verb{\link{build}()}
}
\description{
Helper function to make common adjustments to the spec object built by \verb{\link{adam_spec}()}
for data sets of type 'adsl' to be used with the \verb{\%>\%}.
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

