% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_spec.R
\name{adjust_adsl_factors}
\alias{adjust_adsl_factors}
\title{Adjust factor (levels) from adsl}
\usage{
adjust_adsl_factors(spec, fctrs, entry = "adsl")
}
\arguments{
\item{spec}{object of class \code{martini_spec}}

\item{fctrs}{named list (column in data) of named vectors with
factor levels (values) and labels (names)}

\item{entry}{name of spec entry to modify, defaults to "adsl"}
}
\value{
the \code{spec} object with all valid modifications to the factor
definitions applied.
The modification of a factor definition is skipped (with info)
if not all factor levels of this factor that were derived from the data set
are included in the modified list of levels.
Additional factor level can be introduced, the user is informed.
}
\description{
Adjust factor (levels) from adsl
}
