% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_adsl.R
\name{adjust_adsl_select}
\alias{adjust_adsl_select}
\title{Adjust column selection from data sets of type 'adsl' in spec object}
\usage{
adjust_adsl_select(spec, add = NULL, drop = NULL, select = NULL, id = "adsl")
}
\arguments{
\item{spec}{spec object to modify}

\item{add, drop}{character vector of columns to add to/discard from
the automated selection that is stored in the \code{select} entry of the
corresponding \code{spec} entry}

\item{select}{character vector of column names to be selected. if not NULL
(the default), arguments \code{add} and \code{drop} will be ignored.
Overrides \code{select} entry of the corresponding \code{spec} entry.}

\item{id}{name of list element to modify in the spec, defaults to "adsl"}
}
\value{
A modified version of \code{spec} to be used as input to \verb{\link{build}()}
}
\description{
Helper function to make common adjustments to the spec object built by \verb{\link{adam_spec}()}
for data sets of type 'adsl' to be used with the \verb{\%>\%}.
}
\details{
if data is provided, the column names stored in the \code{select} slot will be
intersected with actually existing column names. User will be informed in case of
misspecifications.
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

