% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_spec.R
\name{adjust_spec}
\alias{adjust_spec}
\title{Adjust spec object}
\usage{
adjust_spec(spec, entry, ...)
}
\arguments{
\item{spec}{spec object to modify}

\item{entry}{name of list element to modify in the spec}

\item{...}{modifications to the \code{spec[[id]]} of the form \verb{<name> = <value>}.}
}
\value{
A modified version of \code{spec} to be used as input to \verb{\link{build}()}
}
\description{
Helper function to make adjustments to the spec object built by
\verb{\link{adam_spec}()} to be used with the pipe.
For example, update of entries for
\code{param}, \code{label}, \code{dupl_ctrl} (bds)
\code{fct_levels}, \code{spec_id}, \code{trt}, \code{id}, (adsl)
\code{valuen}, \code{value}, \code{count} (occds)
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

