% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_adsl.R
\name{adsl_dict}
\alias{adsl_dict}
\title{create adsl dictionary from column labels/names}
\usage{
adsl_dict(adsl, param = "param", label = "label")
}
\arguments{
\item{adsl}{adsl-like data set}

\item{param, label}{column name for resulting tibble where column name
(param) and label are stored}
}
\value{
a \code{tibble}
}
\description{
create adsl dictionary from column labels/names
}
