% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_adsl.R
\name{adsl_identify}
\alias{adsl_identify}
\alias{adsl_identify_dttm}
\alias{adsl_identify_janitor}
\alias{adsl_identify_combined}
\alias{adsl_identify_redundants}
\alias{adsl_identify_flags}
\alias{adsl_identify_factors}
\title{identify/categorize columns from adsl}
\usage{
adsl_identify_dttm(adsl)

adsl_identify_janitor(adsl)

adsl_identify_combined(adsl, dict, dict_label = "label", dict_param = "param")

adsl_identify_redundants(adsl, id, trt, clmn_flag)

adsl_identify_flags(adsl, dict, dict_param = "param", dict_label = "label")

adsl_identify_factors(
  adsl,
  id,
  clmn_flag,
  dict,
  dict_param = "param",
  dict_label = "label"
)
}
\arguments{
\item{adsl}{data set in which to identify date time columns}

\item{dict}{tibble with column names and labels}

\item{dict_label, dict_param}{column names in dict for the label and column
name, resp., defaulting to \code{label} and \code{param}, resp.}

\item{id, trt}{user-selected column names for ID and treatment column}

\item{clmn_flag}{(redundants only) character vector of names identified as flags}
}
\value{
character vector of column names in \code{adsl} that were
identified as candidates for a given category
}
\description{
family of helper functions to identify
}
\details{
Columns meeting the following criteria are returned

\code{adsl_identify_dttm()}:  \code{assertive.types::is_date()} is TRUE,
the label contains strings 'year', 'month', 'day', 'date' or 'time'
(not case sensitive), class is one of 'difftime',
'hms', 'Period', 'POSIXct', 'POSIXt', 'Date'

\code{adsl_identify_janitor()}: identification via
\code{janitor::remove_empty(which = 'cols')}, \code{janitor::remove_constant(na.rm = TRUE)}

\code{adsl_identify_redundants()}: redundant columns to selected trt and id columns
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

