% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{one stop shop for building the data set
Allows to
{\itemize
\item build from spec (with or without data already attached)
\item build from ads path (spec is created and used to build)
\item create spec only (also available from \code{adam_spec()})}}
\usage{
build(
  spec = NULL,
  path = NULL,
  spec_only = FALSE,
  filter = NULL,
  keep = NULL,
  drop = NULL,
  join = dplyr::inner_join,
  attach_data = FALSE
)
}
\arguments{
\item{spec}{}

\item{path}{the path to the ads files}

\item{spec_only}{if build from path, don't apply the just created spec to data set}

\item{filter}{a character vector of conditions to be passed to \code{dplyr::filter()}, e.g. regarding visits, treatment arms or parameters. Defaults to NULL.}

\item{keep}{character vector defining the subset of data sets in the given \code{path} to create the specification for (e.g. \code{c('adsl', 'advs'))}).
If both \code{keep} and \code{drop} are specified, \code{keep} overrides \code{drop}. Defaults to NULL.}

\item{drop}{character vector defining a subset of data sets in the given \code{path} to be excluded from the list of specifications (e.g. \code{'adqseq5d')}). Defaults to NULL.}

\item{join}{either function to join data sets (e.g. \code{dplyr::full_join} or a character (vector) giving the names of the data sets containing the .ids to keep (e.g. \code{join= c('adxb', 'adlb')}). defaults to \code{dplyr::inner_join}}

\item{attach_data}{boolean. attach the imported raw data}
}
\value{
\code{adam_spec()} returns named list of specifications that can be passed to the \code{adam_prep()} function.
Each element contains the specification for a single data set and is named with the domain abbreviation (e.g. adsl, adqskccq).
The list can be manually adjusted if required, e.g. adding further specifications or altering existing ones.
}
\description{
\code{adam_spec()} matches file names in the given path against an internal library to decide on which \code{adam_*_spec()} function to use for which data set.
Only files in the library will be processed, the rest will be ignored. Names of unprocessed files will be printed to the console.
For those, specifications may be created manually using the appropriate \code{adam_spec_*()} function and appended to the specification list created by \code{adam_*_spec()}.

Individual filters are only applied if the resulting data set has a positive number of rows (ignoring those causing errors or yielding a 0-row data set).

Please refer to the documentations of the \code{adam_spec_*()} functions for full details.
}
\seealso{
\code{\link{adam_spec_adsl()}}, \code{\link{adam_spec_bds()}}
}
