% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_out_tte.R
\name{build_out_tte}
\alias{build_out_tte}
\title{Prepare outcome from adtte for MARTINI}
\usage{
build_out_tte(
  data = NULL,
  file = NULL,
  filter = NULL,
  cut = NULL,
  unit = "unit",
  label = ".out",
  columns = NULL
)
}
\arguments{
\item{data}{adtte-like tibble, e.g. output of haven::read_sas('adtte.sas7bdat')}

\item{file}{path to ADaM-like tte data set, e.g. files/adtte.sas7bdat, ignored if data is provided}

\item{filter}{character vector to be applied to the data set (e.g. to select PARAMCD or subset to particular population)}

\item{cut}{optional numeric, required for binarized version from tte data. Note that cut must provided in the same unit as the time column (e.g. AVAL)}

\item{unit}{only relevant if outcome should be binarized, i.e. if \code{cut} is provided.
\code{unit} should describe the unit of the numeric (time) column in data set ass well as of \code{cut}, and will be included in outcome values}

\item{label}{optional label for the binarized outcome variable, defaults to '.out'}

\item{columns}{list defining the mapping of required information to columns in the data with default entries. See details for defaults.}
}
\value{
A tibble with column \code{.id} and either an additional character column \code{.out} for the binarized version
or with the addition of the pair \code{.time} and \code{.status} for tte outcomes.

Observations with missing values in either \code{.time} or \code{.status} will be removed.
In this case, a notification with the number of removed observations will be printed to the console.

Note that sample sizes may differ for binarization and tte outcome, as subjects are dropped if the observed time is censored and below \code{cut}.
}
\description{
Convenience function to prepare an outcome object from adtte-like data sets, either for tte or binarized endpoint.
}
\details{
Note that \code{unit} is solely a description to be included in the outcome value (e.g. "event in first 2 year(s)").
No conversion of the time is done to \code{unit}.

Column mapping defaults are chosen to match ADaM adtte data sets:
\describe{
\item{\code{id}}{\code{'USUBJID'}}
\item{\code{time}}{\code{'AVAL'}}
\item{\code{censor}}{\code{'CNSR'}}
}
If \code{status} instead of \code{censoring} indicator should be used,
specify e.g. status = 'event' (defaults to \code{NULL}).
If both are defined, \code{status} is used, \code{censor} is ignored.
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

