% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_spec.R
\name{check_adjust}
\alias{check_adjust}
\title{check key value pair inputs for adjust_* functions}
\usage{
check_adjust(spec, entry, modifications)
}
\arguments{
\item{spec}{spec object to modify}

\item{entry}{name of list element to modify in \code{spec}}

\item{modifications}{list of key value pairs defining adjustments}
}
\value{
subset of modifications that are valid to apply
}
\description{
check key value pair inputs for adjust_* functions
}
