% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_adsl_select.R
\name{check_adjust_adsl_select}
\alias{check_adjust_adsl_select}
\title{checks for adjustments in \code{adjust_adsl_select()}}
\usage{
check_adjust_adsl_select(spec, add, drop, select, entry = "adsl")
}
\arguments{
\item{spec}{object of class \code{martini_spec}}

\item{add, drop}{character vectors of columns to add to/discard from current selection}

\item{select}{character vector of columns to select (override current selection)}

\item{entry}{name of spec entry to modify}
}
\value{
named list of valid modifications to apply
(entries \code{add}, \code{drop}, \code{select})
}
\description{
checks for adjustments in \code{adjust_adsl_select()}
}
