% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_spec.R
\name{check_adjust_factors}
\alias{check_adjust_factors}
\title{Helper for factor (level) adjustment of spec}
\usage{
check_adjust_factors(spec, fctrs, entry = "adsl")
}
\arguments{
\item{spec}{object of class martini_spec}

\item{fctrs}{named list(column in data) of named vectors with
factor levels (values) and labels (names)}

\item{entry}{name of spec entry to modify, defaults to 'adsl'}
}
\value{
subset of \code{fctrs} that are valid to apply, \code{NULL} if none
}
\description{
Helper for factor (level) adjustment of spec
}
\details{
The function checks if the modifications to the factor definitions
provided in \code{fctrs} are valid:
\itemize{
\item The names of \code{fctrs} top level entries must be present in the data
of the entry (checked by attached data or dictionary).
Else: ignore with message
\item The factor levels provided must contain all current/known levels.
Else: ignore with warning.
\item New factor levels (not seen in data) can be introduced (with message)
\item if only levels are provided, labels are created via \code{purrr::set_names()}
}
}
