% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_helpers.R
\name{check_and_guess_column}
\alias{check_and_guess_column}
\title{Check role specification for ADaM data set}
\usage{
check_and_guess_column(
  data,
  role,
  column_spec = NULL,
  type = c("adsl", "bds", "occds"),
  spec_id = NULL,
  required = TRUE,
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{data set to check}

\item{role}{character. the role to check, e.g. "param", "id", "value" or
"time"}

\item{column_spec}{character. the selected column name. will be for
presence in \code{data}and type. if NULL (the default), it will be guessed based
on \code{domain} or \code{type}.}

\item{type}{character. either "bds" or "occds"}

\item{spec_id}{character. an optional id for the specification that is used
for informative warnings}

\item{required}{boolean. \code{TRUE}, if \code{role} is required, \code{FALSE} if optional.}

\item{call}{the execution environment of a currently running function.}
}
\value{
A list with \code{role}, the column name \code{column} or \code{NULL} (if column check was
not successful or no column for \code{role} could be guessed), \code{required} (passed
as-is from the input) and a boolean \code{check_passed}, indicating, if all checks
on the column have passed. Throws an informative warning if any check fails.
}
\description{
Checks, if provided column for a role in an ADaM data set is present in the
data. If no column is provided, it is guessed based on ADaM standards.
}
