% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_feature.R
\name{check_count}
\alias{check_count}
\title{Check for variables that resemble count variables}
\usage{
check_count(x, thres = NULL, non_neg = TRUE, quiet = FALSE)
}
\arguments{
\item{x}{data set to check}

\item{thres}{number of distinct integer values}

\item{non_neg}{logical controlling whether to only consider variables
with non-negative values. Defaults to TRUE.}

\item{quiet}{whether to suppress printing messages to the console.
defaults to FALSE.}
}
\value{
invisibly returns a list for downstream use in \code{\link[=check_feature]{check_feature()}}
}
\description{
Identify variables that only have only (non-negative) integer values and a
relatively small number of distinct values
}
