% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_feature.R
\name{check_feature}
\alias{check_feature}
\title{check feature matrix}
\usage{
check_feature(
  x,
  check_low_freq = TRUE,
  check_other = TRUE,
  check_missing = TRUE,
  check_count = TRUE,
  quiet = TRUE,
  verbose = TRUE,
  thres_count = 30,
  thres_low_freq = NULL,
  thres_missing = NULL,
  ...
)
}
\arguments{
\item{x}{feature matrix to check, such as the output of \code{\link[=build]{build()}}.}

\item{check_low_freq, check_other, check_missing, check_count}{logicals to control which checks to include. All default to TRUE.}

\item{quiet}{logical controlling whether any informative messages are
printed to the console}

\item{verbose}{TRUE by default, FALSE results in messages used for
\code{\link[=prepare_ml]{prepare_ml()}}}

\item{thres_count}{\code{\link[=check_count]{check_count()}}}

\item{thres_low_freq}{passed to \code{\link[=check_freq]{check_freq()}}}

\item{thres_missing}{passed to \code{\link[=check_non_missing]{check_non_missing()}}}

\item{...}{arguments to be passed to other methods}
}
\value{
a nested list, where each entry has a \code{vars} entry that contains
column names that do not pass the respective check (if check is included),
character of length 0 for empty sets.
NULL if not tested at all.
}
\description{
Running \code{check_feature()} is by default run in \code{\link[=prepare_ml]{prepare_ml()}} on the
input \code{feature} to notify the user on sources of potential issues.
}
