% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filter.R
\name{check_filter}
\alias{check_filter}
\title{Check filter}
\usage{
check_filter(data, filter)
}
\arguments{
\item{data}{data set that the given filter(s) should be applied to}

\item{filter}{an expression to be used in \code{dplyr::filter()}, (either a single filter, or multiple ones separated by ',')}
}
\value{
The function returns a logical of length equal to the number of individual filters present in the \code{filter} argument.
TRUE indicates that the filter applied individually to \code{data} returns a data set with positive number of rows.
On the other hand, FALSE indicates either an empty result after filtering or a non-applicable filter
(e.g. if the column used in the filter definition is not present in \code{data}).
}
\description{
Check a given (set of) filter(s) against a data set and assess whether or not a non-empty data set would be returned.
}
\section{Authors}{


Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

