% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_feature.R
\name{check_freq}
\alias{check_freq}
\title{Identify factors with low frequency classes}
\usage{
check_freq(x, thres = NULL, quiet = FALSE)
}
\arguments{
\item{x}{data set to check}

\item{thres}{integer.
Factors with at least one class of size smaller than \code{thres} will be identified.
Defaults to NULL,}

\item{quiet}{whether to suppress printing messages to the console.
defaults to FALSE.}
}
\value{
invisibly returns a list for downstream use in \code{\link[=check_feature]{check_feature()}}
}
\description{
Identify factors with low frequency classes
}
\details{
If \code{x} is the output of \verb{\link{prepare_ml}()}, the tibble \code{x$data$prep$train} is checked for
factors with at least one low frequency class.
Please refer to the package vignette for further details.
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

