% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_freq.R
\name{check_non_missing}
\alias{check_non_missing}
\title{Check for the proportion of non-missing values}
\usage{
check_non_missing(x, thres = NULL, quiet = FALSE)
}
\arguments{
\item{x}{data set to check}

\item{thres}{Minimum proportion of data available}

\item{quiet}{whether to suppress printing messages to the console.
defaults to FALSE.}
}
\value{
invisibly returns a list for downstream use in \code{\link[=check_feature]{check_feature()}}
}
\description{
Check for the proportion of non-missing values
}
