% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_feature.R
\name{check_nzv}
\alias{check_nzv}
\title{Check (near) zero variance}
\usage{
check_nzv(x, thres_freq = NULL, thres_unique = NULL, quiet = FALSE)
}
\arguments{
\item{x}{data set to check}

\item{thres_freq, thres_unique}{by default (\code{NULL}), the respective
default of \code{prepare_ml()} is used}

\item{quiet}{whether to suppress printing messages to the console.
defaults to FALSE.}
}
\value{
invisibly returns a list for downstream use in \code{\link[=check_feature]{check_feature()}}
}
\description{
Check (near) zero variance
}
