% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_helpers.R
\name{check_occds_occur}
\alias{check_occds_occur}
\title{Check for non-constant --OCCUR column in occds}
\usage{
check_occds_occur(data, domain, filters = NULL)
}
\arguments{
\item{data}{data set to check}

\item{domain}{name of data set}

\item{filters}{filters to be applied before checking for issues.
defaults to NULL, in which case no filters are applied.}
}
\value{
nothing is returned if no potential issues were detected.
If domain is adcm or admh, invisibly returns the suggested filter.
}
\description{
Check for non-constant --OCCUR column in occds
}
