% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_feature.R
\name{check_other_class}
\alias{check_other_class}
\title{Check for occurrence of level that would cause issue with lumping}
\usage{
check_other_class(x, other2_class = NULL, quiet = FALSE)
}
\arguments{
\item{x}{data set to check}

\item{other2_class}{name of class to check for.
If \code{NULL} (the default), uses the default of \code{\link[=step_other2]{step_other2()}}'s argument
\code{other}.}

\item{quiet}{whether to suppress printing messages to the console.
defaults to FALSE.}
}
\value{
invisibly returns a list for downstream use in \code{\link[=check_feature]{check_feature()}}
}
\description{
Check for occurrence of level that would cause issue with lumping
}
