% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_helpers.R
\name{check_role}
\alias{check_role}
\title{Check role specification for ADaM data set}
\usage{
check_role(
  data,
  role,
  column_spec = NULL,
  type = c("bds", "occds"),
  spec_id = NULL,
  required = TRUE,
  call = rlang::caller_env()
)
}
\arguments{
\item{column_spec}{character. the selected column name. will be for
presence in \code{data}and type. if NULL (the default), it will be guessed based
on \code{domain} or \code{type}.}

\item{type}{character. either "bds" or "occds"}

\item{required}{boolean. \code{TRUE}, if \code{role} is required, \code{FALSE} if optional.}

\item{call}{the execution environment of a currently running function.}

\item{colnames_data}{character vector. column names of the data set to check
against
role character. the .g. "param", "id", "value" or
"time"}

\item{domain}{character. an optional id for the specification that is used
for informative warnings}
}
\value{
A list with \code{role}, the column name \code{column} or \code{NULL} (if column check was
not successful or \code{role} could not be guessed) and a boolean \code{check_pass},
indicating, if all checks on the column have passed.
Throws an informative warning if any of the checks fails.
}
\description{
Checks, if provided column for a role in an ADaM data set is present in the
data and of the correct type. If no column is provided, it is guessed based
on ADaM standards.
}
