% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corr_stretch}
\alias{corr_stretch}
\title{Stretch a correlation matrix to long format}
\usage{
corr_stretch(x, shave = FALSE)
}
\arguments{
\item{x}{a symmetric matrix containing the correlations}

\item{shave}{logical. if TRUE, only the lower triangle of the correlation
matrix is kept.}
}
\value{
A tibble with the pair of variable names in the columns \code{x} and \code{y} and the
corresponding correlation in the column \code{r}
}
\description{
Stretch a correlation matrix from \code{\link[stats:cor]{stats::cor()}} to a long format tibble
}
