% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{corrr_mini}
\alias{corrr_mini}
\title{Correlation matrix in long format}
\usage{
corrr_mini(x, method = "pearson", use = "pairwise.complete.obs", shave = FALSE)
}
\arguments{
\item{x}{A tibble, data frame or matrix containing numeric columns to be correlated.}

\item{method, use}{Arguments that are passed to \code{\link[stats:cor]{stats::cor()}}}

\item{shave}{logical. if TRUE, only the lower triangle of the correlation matrix is kept.}
}
\value{
A tibble with the pair of variable names in the columns \code{x} and \code{y} and the
corresponding correlation in the column \code{r}

TODO deprecate
}
\description{
Calculate correlation matrix of a numeric data set and stretch to long format for convenient filtering..
}
