% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_helpers.R
\name{create_dict}
\alias{create_dict}
\title{Title}
\usage{
create_dict(spec_entry)
}
\arguments{
\item{spec_entry}{Top level entry of an object of class \code{martini_spec}.}
}
\value{
the dictionary
}
\description{
Title
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
TODO complete docu
}

