% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_helpers.R
\name{create_spec_out}
\alias{create_spec_out}
\title{Create output object for build specifications}
\usage{
create_spec_out(..., type = c("adsl", "bds", "occds"), attach_data = TRUE)
}
\arguments{
\item{...}{output objects}

\item{type}{character. either "bds" or "occds"}

\item{attach_data}{boolean indicating whether the imported raw data is included in
the output. Defaults to FALSE.}
}
\value{
Output object of \verb{adam_spec_*()}
}
\description{
Create output object for build specifications
}
\seealso{
\code{\link[=adam_spec_bds]{adam_spec_bds()}}
\code{\link[=adam_spec_occds]{adam_spec_occds()}}
}
