% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fct_na_to_level}
\alias{fct_na_to_level}
\title{create alias for fct_explicit_na and fct_na_value_to_level based on
available forcats version}
\usage{
fct_na_to_level(f, level)
}
\arguments{
\item{f}{A factor (or character vector)}

\item{level}{Level to use for missing values: this is what NAs will be changed to}
}
\value{
a factor
}
\description{
create alias for fct_explicit_na and fct_na_value_to_level based on
available forcats version
}
