% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_default}
\alias{get_default}
\title{get fct defaults}
\usage{
get_default(fun, arg = NULL, unlist = TRUE, unname = TRUE)
}
\arguments{
\item{fun}{function name}

\item{arg}{character (vector) of names of arguments to show default values for (exact match required)}

\item{unlist, unname}{logicals}
}
\value{
depending on values unlist and unname either a list or a vector
}
\description{
get fct defaults
}
