% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_bds.R
\name{import_info}
\alias{import_info}
\title{Import file and collect info}
\usage{
import_info(file)
}
\arguments{
\item{file}{filepath}
}
\value{
list containing data and corresponding md5 sum and file size
}
\description{
Import file and collect info
}
