% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info_filter.R
\name{info_filter}
\alias{info_filter}
\title{Extract filter info from a spec object}
\usage{
info_filter(spec, filter = NULL, quiet = FALSE)
}
\arguments{
\item{spec}{spec object as returned by \code{\link{adam_spec}()}}

\item{filter}{if not \code{NULL} (default), applied filters are
compared against a reference set,
identifying filters that cannot be applied to the data without an error}

\item{quiet}{if \code{TRUE} instead of printing message to console,
return list with messages on applied and discarded filters.
Defaults to \code{FALSE}.}
}
\value{
List of applied filters by data set is printed to the console.
If \code{filter} is applied, missing filters are listed, if any are identified.
}
\description{
Extract applied filters and (optionally) compare to reference set
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

