% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/martini-package.R
\docType{package}
\name{martini-package}
\alias{martini}
\alias{martini-package}
\title{martini: BMDI MLAI pipeline - Prepare Module}
\description{
Prepare module of the BMDI MLAI pipeline. It contains functions to conveniently transform a set of ADaM data sets (adsl, bds and occds) into a data object that can be used for machine learning, e.g. using the BMDI MLAI pipeline.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://by-gitlab.de.bayer.cnb/EQYKG/batcave/-/issues}
}

}
\author{
\strong{Maintainer}: Ahrens Maike \email{maike.ahrens.ext@bayer.com} (ahrensmaike)

Authors:
\itemize{
  \item Voss Sebastian \email{sebastian.voss.ext@bayer.com} (svoss09)
}

}
\keyword{internal}
