% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-martini_feat.R
\docType{data}
\name{martini_feat}
\alias{martini_feat}
\title{Example feature matrix.}
\format{
A tibble with one row per subject and the features in the column. See the documentation of
\code{\link{build}()} for details on the object structure.
}
\source{
Simulated data
}
\usage{
martini_feat
}
\description{
Feature matrix of the included example study in \code{system.file("ads", package = "martini")}
created by \code{\link{build}()} from \code{\link{martini_spec}}.
}
\keyword{datasets}
