% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-martini_ml.R
\docType{data}
\name{martini_ml_class}
\alias{martini_ml_class}
\title{Example ML classification object.}
\format{
A \code{martini_ml} object. See the documentation of \code{\link[=prepare_ml]{prepare_ml()}}
for further details on the object structure.
}
\source{
Simulated data
}
\usage{
martini_ml_class
}
\description{
ML object of the example study in \code{system.file("ads", package = "martini")} and the classification
outcome \link{martini_outc_class} as created by \code{\link[=prepare_ml]{prepare_ml()}}.
}
\keyword{datasets}
