% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-martini_ml.R
\docType{data}
\name{martini_ml_surv}
\alias{martini_ml_surv}
\title{Example ML survival object.}
\format{
A \code{martini_ml} list object. See the documentation of \code{\link[=prepare_ml]{prepare_ml()}} for further details on the object structure.
}
\source{
Simulated data
}
\usage{
martini_ml_surv
}
\description{
ML object of the example study in \code{system.file("ads", package = "martini")} and the survival
outcome \link{martini_outc_surv} as created by \code{\link[=prepare_ml]{prepare_ml()}}.
}
\keyword{datasets}
