% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-martini_outc_class.R
\docType{data}
\name{martini_outc_class}
\alias{martini_outc_class}
\title{Example classification outcome}
\format{
A tibble with one row per subject and the columns \code{.id} and \code{.out}.
}
\source{
Simulated data
}
\usage{
martini_outc_class
}
\description{
Classification outcome for the included example study in \code{system.file("ads", package = "martini")}.
}
\keyword{datasets}
