% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-martini_outc_regr.R
\docType{data}
\name{martini_outc_regr}
\alias{martini_outc_regr}
\title{Example regression outcome}
\format{
A tibble with one row per subject and the columns \code{.id} and \code{.out}.
}
\source{
Simulated data
}
\usage{
martini_outc_regr
}
\description{
Regression outcome for the included example study in \code{system.file("ads", package = "martini")}.
}
\keyword{datasets}
