% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-martini_outc_surv.R
\docType{data}
\name{martini_outc_surv}
\alias{martini_outc_surv}
\title{Example survival outcome}
\format{
A tibble with one row per subject and the columns \code{.id}, \code{.time} and \code{.status}.
}
\source{
Simulated data
}
\usage{
martini_outc_surv
}
\description{
Survival outcome for the included example study in \code{system.file("ads", package = "martini")}.
}
\keyword{datasets}
