% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_corr_keep.R
\name{martini_recipes_remove_cols}
\alias{martini_recipes_remove_cols}
\title{Removes columns if options apply}
\usage{
martini_recipes_remove_cols(new_data, object, col_names = character())
}
\arguments{
\item{new_data}{A tibble.}

\item{object}{A step object.}

\item{col_names}{A character vector, denoting columns to remove. Will
overwrite \code{object$removals} if set.}
}
\value{
\code{new_data} with column names removed if specified by \code{col_names} or
\code{object$removals}.
}
\description{
This helper function removes columns based on character vectors.
}
\author{
this function is basically a copy of recipes::recipes_remove_cols()
}
