% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-martini_spec.R
\docType{data}
\name{martini_spec}
\alias{martini_spec}
\title{Example specification object.}
\format{
A list with the entries \code{adsl}, \code{adlb}, \code{admh} and \code{advs}.
See the documentation of
\code{\link[=adam_spec]{adam_spec()}} for further details on the object structure.
}
\source{
Simulated data
}
\usage{
martini_spec
}
\description{
Specification object of the included example study in
\code{system.file("martini_example_study/ads", package = "martini")}
as created by \code{\link[=adam_spec]{adam_spec()}}.
}
\keyword{datasets}
