% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bds.R
\name{pivot_prepare_bds}
\alias{pivot_prepare_bds}
\title{Prepare bds data for pivoting step in build}
\usage{
pivot_prepare_bds(
  bds_full,
  spec,
  values_fn = NULL,
  arrange = NULL,
  clean_fn = ~stringr::str_replace_all(.x, "[:punct:]|[:space:]", "_"),
  names_sep = "_"
)
}
\arguments{
\item{bds_full}{original bds-type data set}

\item{values_fn, names_sep}{simply written to the \code{pivot_args} output for the sake of completeness}

\item{arrange}{}

\item{clean_fn}{}

\item{filter}{subsetting}

\item{id}{additional columns to reduce data set to actual columns required for pivoting}
}
\value{
A list containing the pivot_wider arguments (pivot_args) as well as the function
to clean column names (clean_fn). The \code{pivot_args} list includes the prepared data set (filtered, arranged)
as well as pivot_wider params (key(s), value, values_fn, names_sep)
}
\description{
Preparation of dataset bds_full as well as parameters to be passed to pivot_wider
in build_bds to allow for appropriate unit testing
}
\details{
Data preparation of bds_full for pivoting includes filtering and arranging
the data set before relevant columns are selected and renamed using \code{clean_fn} (\code{param}, \code{time} only)
If the prepared data set has more than one level in the \code{time} column,
names_from will be a vector of the form \code{c(param, time)}
}
