% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_helpers.R
\name{prepare_col_selection}
\alias{prepare_col_selection}
\title{Prepare column selection}
\usage{
prepare_col_selection(
  data,
  ...,
  type = c("adsl", "bds", "occds"),
  call = rlang::caller_env()
)
}
\arguments{
\item{data}{data set to check}

\item{...}{objects containing the column names for the roles in the data set
following the standard naming convention in the \verb{adam_spec_*()} functions
(\code{id}, \code{value}, etc.).}

\item{type}{character. either "bds" or "occds"}

\item{call}{the execution environment of a currently running function.}
}
\value{
A list with \code{col_select} containing the column names for the roles in the
data set and \code{use_for_build} indicating, if all checks on the columns have
passed.
}
\description{
Prepare column selection in \verb{adam_spec_*()} functions for output. This
includes checks for column presence and guessing of column names based on
ADaM standards and transformation into a standard format for further use
within the \verb{adam_spec_*()} functions.
}
