% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ml.R
\name{prepare_ml}
\alias{prepare_ml}
\title{Prepare ML ready data set from outcome and predictor data}
\usage{
prepare_ml(
  feature,
  outcome,
  outcome_name = NULL,
  level_order = NULL,
  prep_recipe = NULL,
  train_prop = 3/4,
  seed = NULL,
  prep_step_normalize = TRUE,
  prep_step_knnimpute = TRUE,
  prep_step_log = TRUE,
  prep_step_corr = TRUE,
  prep_step_dummy = TRUE,
  thres_log = 2,
  thres_count = 10,
  thres_corr = 0.9,
  thres_lump = 0.05,
  thres_imp = 0.8,
  thres_nzv_freq = 95/5,
  thres_nzv_unique = 10,
  vars_imp_ignore = NULL,
  vars_fct_expl_na = NULL,
  vars_ordinalscore = NULL,
  one_hot = TRUE,
  log_base = exp(1),
  outlier_remove = FALSE,
  outlier_ctrl = list(coef = 3),
  quiet = FALSE
)
}
\arguments{
\item{feature}{feature matrix in wide format, e.g. output object of \code{build()}, i.e. containing \code{.id} column and predictors}

\item{outcome}{tibble containing \code{.id} column and the outcome of interest}

\item{outcome_name}{if NULL (default), the first column that's not .id is chosen for outcome_name and the outcome_type is guessed to be either classification or regression.
One may also provide a single character giving the name of the outcome column OR a vector of length two giving the column names for the 'time' and 'status' data in survival analysis,
where .time is numeric and .status is binary with 0 coding for censored, and 1 coding for event. Currently, only right-censoring is supported. Please note, that survival will never be guessed}

\item{level_order}{= NULL (only used for classification)}

\item{prep_recipe}{= NULL,}

\item{train_prop}{= 3/4}

\item{seed}{= NULL,}

\item{prep_step_normalize}{= TRUE,}

\item{prep_step_knnimpute}{= TRUE,}

\item{prep_step_log}{= TRUE,}

\item{prep_step_corr}{= TRUE,}

\item{prep_step_dummy}{FALSE converted  variables to be}

\item{thres_log}{= 2,}

\item{thres_count}{= 10}

\item{thres_corr}{= .9,}

\item{thres_lump}{= 0.05}

\item{thres_imp}{= 0.8}

\item{thres_nzv_freq}{=95/5}

\item{thres_nzv_unique}{= 10}

\item{vars_imp_ignore}{= NULL}

\item{vars_fct_expl_na}{= NULL}

\item{vars_ordinalscore}{= NULL, convert ordinal factor variables into numeric scores}

\item{one_hot}{= TRUE}

\item{log_base}{= exp(1),}

\item{outlier_remove}{= FALSE,}

\item{outlier_ctrl}{= list(coef = 3)}

\item{quiet}{=FALSE

'  ='_',}
}
\description{
Prepare ML ready data set from outcome and predictor data
}
