% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ml.R
\name{prepare_ml_data_split}
\alias{prepare_ml_data_split}
\title{split data in train and test}
\usage{
prepare_ml_data_split(data, train_prop, strata_trt, seed = NULL, outcome_mode)
}
\arguments{
\item{data}{data to split}

\item{train_prop}{proportion to use for training split, must be in (0.5, 1]}

\item{strata_trt}{logical}

\item{seed}{defaults to NULL}

\item{outcome_mode}{used in stratification}
}
\value{
a named list containing
}
\description{
split data in train and test
}
