% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ml_recipe.R
\name{prepare_ml_recipe}
\alias{prepare_ml_recipe}
\title{create recipe from data}
\usage{
prepare_ml_recipe(
  data,
  prep_recipe = NULL,
  corr_method = "pearson",
  corr_use = "pairwise.complete.obs",
  thres_list = NULL,
  step_list = NULL,
  vars_imp_ignore = c(".trt"),
  vars_fct_expl_na = NULL,
  vars_ordinalscore = NULL,
  vars_keep_corr = NULL,
  level_other,
  one_hot,
  log_base
)
}
\arguments{
\item{data}{raw data set to create recipe for}

\item{prep_recipe}{if NULL, recipe will be created}

\item{corr_method, corr_use}{defaulting to corr_method "pearson" and corr_use "pairwise.complete.obs"}

\item{thres_list, step_list}{named list objects collecting all threshold values
and step selection info, resp. please refer to the documentation of the \verb{thres_*} and \verb{prep_step_*} arguments
in \code{\link{prepare_ml}()} for detailed documentation and list entry names.}

\item{vars_imp_ignore, vars_fct_expl_na, vars_ordinalscore, vars_keep_corr}{vars relevant
to steps. see \code{\link{prepare_ml}()} for details.}

\item{level_other}{name for the "other"-category in \code{recipes::step_other()}. defaults to "other".}

\item{one_hot, log_base}{see \code{\link{prepare_ml}()}}
}
\value{
a named list with entries containing
\itemize{
\item the prepared recipe
\item info on steps included in the recipe
\item a list of relevant variables
\item a list of thresholds used
\item \code{high_corr} a tibble listing correlations above \code{thres_corr}. \code{NULL} if \code{step_list$prep_step_corr = FALSE}.
}
}
\description{
create recipe from data
}
\seealso{
\code{\link{prepare_ml}()}
}
