% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ml_vars.R
\name{prepare_replace}
\alias{prepare_replace}
\title{consistent renaming of character vectors/factor levels}
\usage{
<<<<<<< HEAD
prepare_replace(x)
=======
prepare_replace(x = NULL)
>>>>>>> 718adf69a48e691e9ad94351f18dbf17115686bf
}
\arguments{
\item{x}{character vector}
}
\value{
list with the updated x, obtained from call \code{stringr::str_replace_all(x, replacement)},
where replacement is returned as separate same-name list entry
}
\description{
consistent renaming of character vectors/factor levels
}
