% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_zap_empty}
\alias{read_zap_empty}
\title{a convenience function combining \code{\link[haven:read_sas]{haven::read_sas()}} with
\code{\link[haven:zap_empty]{haven::zap_empty()}} functionality}
\usage{
read_zap_empty(data_file, catalog_file = NULL)
}
\arguments{
\item{data_file}{sas7bdat file to read}

\item{catalog_file}{catalog file to read}
}
\description{
a convenience function combining \code{\link[haven:read_sas]{haven::read_sas()}} with
\code{\link[haven:zap_empty]{haven::zap_empty()}} functionality
}
