% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_outcome.R
\name{simulate_outcome}
\alias{simulate_outcome}
\title{Simulate outcome for MLAI pipeline}
\usage{
simulate_outcome(
  X,
  beta = NULL,
  type = c("regression", "classification", "survival"),
  ctrl_regr = list(b0 = 0, sd = 0.4),
  ctrl_class = list(prob_ev = 0.5, mult_beta = 1),
  ctrl_surv = list(surv_mean = 18, cens_mean = 18, cens_max = 36, mult_beta = 1, int =
    FALSE)
)
}
\arguments{
\item{X}{Prepared feature matrix (as tibble or matrix) with standardized and
potentially log-transformed numeric variables and dummy-coded categorical
variables. If interaction effects are desired, a corresponding column has to
be present in \code{X} (e.g. \code{X$interaction_A_B = X$A*X$B}). Needs to
contain an id column named ".id".}

\item{beta}{Named effect vector with the corresponding effects for the
columns in \code{X}. Only non-zero effects have to be specified}

\item{type}{type of the simulated outcome: "regression", "classification" or
"survival"}

\item{ctrl_regr}{list with the settings for regression outcome, simulated by
a linear regression model
\describe{
\item{\code{b0}}{model intercept}
\item{\code{sd}}{standard deviation of the model error term}
}}

\item{ctrl_class}{list with settings for classification outcome
(event vs. no event), simulated by a logistic regression model
\describe{
\item{\code{prob_ev}}{event probability}
\item{\code{mult_beta}}{enhancement factor for the beta coefficient to control
signal-to-noise ratio}
}}

\item{ctrl_surv}{list with settings for right-censored survival outcome,
simulated by a proportional hazard model with time-constant baseline hazard
\describe{
\item{\code{surv_mean}}{mean survival time}
\item{\code{cens_mean}}{mean censoring time, NULL for no censoring}
\item{\code{cens_max}}{max censoring time, NULL for no censoring}
\item{\code{mult_beta}}{enhancement factor for the beta coefficient to control
signal-to-noise ratio}
\item{\code{int}}{boolean, round the survival times to the next highest integer}
}}
}
\value{
A tibble with \code{nrow(X)} rows and columns \code{.id} and \code{.out} for regression or
classification outcome or columns \code{.id}, \code{.time} and \code{.status} for survival
outcome.
}
\description{
Simulate regression, classification or survival outcome based on
feature matrix and effect vector
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

