% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{skw}
\alias{skw}
\title{skewness ####
e1071 was only used for e1071::skewness in prepare_ml_vars,default type = 3}
\usage{
skw(x, na.rm = FALSE)
}
\arguments{
\item{x}{a numeric vector containing the values whose skewness is to be computed.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
The estimated skewness of x.
}
\description{
skewness ####
e1071 was only used for e1071::skewness in prepare_ml_vars,default type = 3
}
