% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_internal.R
\docType{data}
\name{spec_cols_required}
\alias{spec_cols_required}
\title{Required spec entry by data type}
\format{
A list with top level elements for each data type, giving the set of
required entries for that data type.
}
\usage{
spec_cols_required
}
\description{
used internally for "use_for_build" check, as well as in the documentation
to help the user spot issues with spec entries
}
\keyword{datasets}
