% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_corr_keep.R
\name{step_corr_keep}
\alias{step_corr_keep}
\title{High correlation filter}
\usage{
step_corr_keep(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  threshold = 0.9,
  use = "pairwise.complete.obs",
  method = "spearman",
  keep = NULL,
  removals = NULL,
  high_corr = NULL,
  skip = FALSE,
  id = recipes::rand_id("corr_keep")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose variables for this step.
See \code{\link[=selections]{selections()}} for more details.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have
been estimated.}

\item{threshold}{A value for the threshold of absolute correlation values.
The step will try to remove the minimum number of columns so that all the
resulting absolute correlations are less than this value.}

\item{use}{A character string for the \code{use} argument to the \code{\link[stats:cor]{stats::cor()}}
function.}

\item{method}{A character string for the \code{method} argument to the
\code{\link[stats:cor]{stats::cor()}} function, defaults to \code{spearman}}

\item{keep}{A character vector, containing variables that should be kept.
These will be prioritized when selecting a representative from a
variable pair with an absolute correlation greater than \code{threshold}
(see details). If \code{NULL}, this step is equivalent to
\code{\link[recipes:step_corr]{recipes::step_corr()}}.}

\item{removals}{A character string that contains the names of columns that
should be removed. These values are not determined until \code{\link[=prep]{prep()}} is
called.}

\item{high_corr}{A tibble containing all correlations above \code{threshold}.
These values are not determined until \code{\link[=prep]{prep()}} is called.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked by
\code{\link[=bake]{bake()}}? While all operations are baked when \code{\link[=prep]{prep()}} is run, some
operations may not be able to be conducted on new data (e.g. processing the
outcome variable(s)). Care should be taken when using \code{skip = TRUE} as it
may affect the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_corr_keep()} creates a \emph{specification} of a recipe step that will
potentially remove variables that have large absolute correlations with other
variables, keeping just one representative from highly correlated variable
pairs. The choice of representatives can be controlled by providing a set of
variables that should be prioritized.
}
\details{
This step can potentially remove columns from the data set. This may
cause issues for subsequent steps in your recipe if the missing columns are
specifically referenced by name. To avoid this, see the advice in the
\emph{Tips for saving recipes and filtering columns} section of
\link[recipes:selections]{recipes::selections}.

This step attempts to remove variables to keep the largest absolute
correlation between the variables less than \code{threshold}.

The filter tries to prioritize predictors for removal based on the global
affect on the overall correlation structure. If you have two
predictors with an absolute correlation above \code{threshold}, the variable with
the larger average correlation with all other predictors will be removed,
unless it is specified in \code{keep} as a variable the user wants to prioritize.
If the absolute correlation of two variables in \code{keep} exceeds the
\code{threshold}, the variable with the larger average correlation to the other
predictors will be removed and and the user is informed by a message in the
console.

When a column has a single unique value, that column will be excluded from
the correlation analysis. Also, if the data set has sporadic missing values
(and an inappropriate value of \code{use} is chosen), some columns will also be
excluded from the filter.

The arguments \code{use} and \code{method} don't take effect if case weights are used
in the recipe.
}
\section{Tidying}{
When you \code{\link[recipes:tidy.recipe]{tidy()}} this step, a tibble is returned with
columns \code{terms} and \code{id}:

\describe{
\item{terms}{character, the selectors or variables selected to be removed}
\item{id}{character, id of this step}
}
}

\section{Tuning Parameters}{
The \code{threshold} parameter can be tuned.
}

\section{Case weights}{


This step performs an unsupervised operation that can utilize case weights.
As a result, case weights are only used with frequency weights. For more
information, see the documentation in \link[recipes:case_weights]{recipes::case_weights} and the
examples on \code{tidymodels.org}.
}

\examples{
\dontshow{if (rlang::is_installed("MASS")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# create a data set
set.seed(1717)
p <- 5
corrm <- matrix(numeric(p^2), ncol = p, nrow = p)
# variable 2 has a higher average correlation 
# than all other variables
corrm[,2] <- corrm[2,] <- .2
# variable 1 and 2 have high correlation
corrm[1,2] <- corrm[2,1] <- .9
diag(corrm) <- 1
X <- MASS::mvrnorm(n = 100, mu = rep(0, p), Sigma = corrm) \%>\% 
  tibble::as_tibble(.name_repair = ~paste0("V", 1:p))

# apply correlation filter without specifying `keep`
rec_prep <- recipes::recipe(~., data = X) \%>\% 
  step_corr_keep(
    recipes::all_numeric_predictors(),
    threshold = .8
  ) \%>\% 
  recipes::prep()

recipes::bake(rec_prep, new_data = NULL)

# make sure that "V2" is kept
rec_keep_prep <- recipes::recipe(~., data = X) \%>\% 
  step_corr_keep(
    recipes::all_numeric_predictors(),
    threshold = .8,
    keep = "V2"
  ) \%>\% 
  recipes::prep()

recipes::bake(rec_keep_prep, new_data = NULL)

# inspect high correlations
rec_keep_prep$steps[[1]]$high_corr
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[recipes:step_corr]{recipes::step_corr()}}
}
\author{
Modified from \code{\link[recipes:step_corr]{recipes::step_corr()}}.
}
