% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_log_skewness.R
\name{step_log_skewness}
\alias{step_log_skewness}
\title{Logarithmic transformation based on skewness}
\usage{
step_log_skewness(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  skewness = NULL,
  base = exp(1),
  offset = 0,
  columns = NULL,
  skip = FALSE,
  id = recipes::rand_id("log_skewness")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose variables for this step.
See \code{\link[=selections]{selections()}} for more details.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have
been estimated.}

\item{skewness}{Numeric threshold for the skewness. If the skewness of a
variable exceeds this threshold, it will be log-transformed. Otherwise,
it will remain as-is. If \code{NULL}, all selected numeric variables will be
transformed.}

\item{base}{A numeric value for the base.}

\item{offset}{An optional value to add to the data prior to logging (to avoid
\code{log(0)}).}

\item{columns}{A character vector of the variable names that are
log-transformed. This field is a placeholder and will be populated once
\code{\link[=prep]{prep()}} is used.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked by
\code{\link[=bake]{bake()}}? While all operations are baked when \code{\link[=prep]{prep()}} is run, some
operations may not be able to be conducted on new data (e.g. processing the
outcome variable(s)). Care should be taken when using \code{skip = TRUE} as it
may affect the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_log_skewness()} creates a \emph{specification} of a recipe step that will log
transform numeric variables if the skewness exceeds a given threshold.
}
\section{Tidying}{
When you \code{\link[recipes:tidy.recipe]{tidy()}} this step, a tibble is returned with
columns \code{terms}, \code{base} , and \code{id}:

\describe{
\item{terms}{character, the selectors or variables selected}
\item{base}{numeric, value for the base}
\item{id}{character, id of this step}
}
}

\section{Case weights}{


The underlying operation does not allow for case weights.
}

\examples{
set.seed(313)
examples <- matrix(exp(rnorm(40)), ncol = 2)
examples <- as.data.frame(examples)

rec <- recipes::recipe(~ V1 + V2, data = examples)

log_trans <- rec |>
  step_log_skewness(recipes::all_numeric_predictors(), skewness = 1)

log_obj <- recipes::prep(log_trans, training = examples)

transformed_te <- recipes::bake(log_obj, examples)
plot(examples$V1, transformed_te$V1)
plot(examples$V2, transformed_te$V2)
}
\seealso{
\code{\link[recipes:step_log]{recipes::step_log()}}
}
\author{
Modified from \code{\link[recipes:step_log]{recipes::step_log()}}.
}
