% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_log_skewness.R
\name{step_log_skewness_undo}
\alias{step_log_skewness_undo}
\title{Undoing logarithmic transformations based on skewness}
\usage{
step_log_skewness_undo(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  base = NULL,
  offset = NULL,
  columns = NULL,
  skip = FALSE,
  id = recipes::rand_id("log_skewed_undo"),
  id_undo = NULL
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose variables for this step.
See \code{\link[=selections]{selections()}} for more details.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have
been estimated.}

\item{base}{A numeric value for the base.}

\item{offset}{An optional value to add to the data prior to logging (to avoid
\code{log(0)}).}

\item{columns}{A character vector of the variable names that are
log-transformed. This field is a placeholder and will be populated once
\code{\link[=prep]{prep()}} is used.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked by
\code{\link[=bake]{bake()}}? While all operations are baked when \code{\link[=prep]{prep()}} is run, some
operations may not be able to be conducted on new data (e.g. processing the
outcome variable(s)). Care should be taken when using \code{skip = TRUE} as it
may affect the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_log_skewness_undo()} creates a \emph{specification} of a recipe step that
will reverse any log transformations that are done by \code{\link[=step_log_skewness]{step_log_skewness()}}.
}
\section{Case weights}{


The underlying operation does not allow for case weights.
}

\section{Tidying}{
When you \code{\link[recipes:tidy.recipe]{tidy()}} this step, a tibble is returned with
columns \code{terms}, \code{base} , and \code{id}:

\describe{
\item{terms}{character, the selectors or variables selected}
\item{base}{numeric, value for the base}
\item{id}{character, id of this step}
}
}

\seealso{
\code{\link[recipes:step_log]{recipes::step_log()}}
}
\author{
Modified from \code{\link[recipes:step_log]{recipes::step_log()}}.
}
