% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_other2.R
\name{step_other2}
\alias{step_other2}
\title{Collapse infrequent categorical levels}
\usage{
step_other2(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  threshold = 0.05,
  other = "other_ml",
  single_low_level = c("as-is", "rename"),
  objects = NULL,
  skip = FALSE,
  id = recipes::rand_id("other2")
)
}
\arguments{
\item{recipe}{A recipe object. The step will be added to the sequence of
operations for this recipe.}

\item{...}{One or more selector functions to choose variables for this step.
See \code{\link[recipes:selections]{recipes::selections()}} for more details.}

\item{role}{Not used by this step since no new variables are created.}

\item{trained}{A logical to indicate if the quantities for preprocessing have
been estimated.}

\item{threshold}{A numeric value between 0 and 1, or an integer greater or
equal to one.  If less than one, then factor levels with a rate of
occurrence in the training set below \code{threshold} will be pooled to \code{other}.
If greater or equal to one, then this value is treated as a frequency and
factor levels that occur less than \code{threshold} times will be pooled to
\code{other}.}

\item{other}{A single character value for the other category, defaults to
\code{"other_ml"}.}

\item{single_low_level}{character controlling handling of a single low
rate/frequency class. Defaults to 'as-is', where data is unmodified if only
a single level meets the criterion for pooling. This is different from
\code{recipes::step_other()}'s behavior ('rename'), where the low rate/frequency
class would not be pooled with other classes but renamed to \code{other}}

\item{objects}{A list of objects that contain the information to pool
infrequent levels that is determined by \code{\link[recipes:prep]{recipes::prep()}}.}

\item{skip}{A logical. Should the step be skipped when the recipe is baked by
\code{\link[recipes:bake]{recipes::bake()}}? While all operations are baked when \code{\link[recipes:prep]{recipes::prep()}} is run, some
operations may not be able to be conducted on new data (e.g. processing the
outcome variable(s)). Care should be taken when using \code{skip = TRUE} as it
may affect the computations for subsequent operations.}

\item{id}{A character string that is unique to this step to identify it.}
}
\value{
An updated version of \code{recipe} with the new step added to the
sequence of any existing operations.
}
\description{
\code{step_other2()} creates a \emph{specification} of a recipe step that will
potentially pool infrequently occurring values into an \code{"other_ml"} category.
}
\details{
The overall proportion (or total counts) of the categories are computed. The
\code{other} category is used to pool any two or more categorical levels whose individual
proportion (or frequency) in the training set is less than \code{threshold}.

If no pooling is done the data are unmodified (although character data may be
changed to factors based on the value of \code{strings_as_factors} in
\code{\link[recipes:prep]{recipes::prep()}}/\code{\link[recipes:recipe]{recipes::recipe()}}).
Otherwise, a factor is always returned with different factor levels.

If \code{threshold} is less than the largest category proportion, all levels
except for the most frequent are collapsed to the \code{other} level.

If \code{other_ml} is in the list of discarded levels, no error occurs.

If no pooling is done, novel factor levels are converted to missing. If
pooling is needed, they will be placed into the other_ml category.

When data to be processed contains novel levels (i.e., not contained in the
training set), the other category is assigned.
}
\section{Differences to step_other()}{
\itemize{
\item a single class subject to lumping is kept as-is as opposed to renamed
to \code{other} (during prep)
\item If the level defined in \code{other} is an original class level that was not
subject to pooling,
the user is informed, but no error is raised.
}
}

\section{TODO check message in test case}{
\itemize{
\item novel factor levels are not pooled with an existing lumped category (during bake)
}
}

\section{Tidying}{
When you \code{\link[recipes:tidy.recipe]{recipes::tidy()}} this step,
a tibble is returned with
columns \code{terms}, \code{retained} , and \code{id}:

\describe{
\item{terms}{character, the selectors or variables selected}
\item{retained}{character, factor levels not pulled into \code{other}}
\item{id}{character, id of this step}
}
}

\section{Case weights}{


This step performs an unsupervised operation that can utilize case weights.
As a result, case weights are only used with frequency weights. For more
information, see the documentation in \link[recipes:case_weights]{recipes::case_weights} and the
examples on \code{tidymodels.org}.



The underlying operation does not allow for case weights.
}

\author{
This step is based on
\href{https://github.com/tidymodels/recipes/blob/d269758cb171698f38f376fcd711de941840657f/R/other.R#L1}{\code{recipes::step_other()}}
with only minor modifications.
}
