% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bds.R
\name{values_fn_default}
\alias{values_fn_default}
\title{Default duplicate handling}
\usage{
values_fn_default(x)
}
\arguments{
\item{x}{vector to be summarized into a single value of the same type}
}
\value{
one-dimensional vector of the same type as \code{x}
}
\description{
Default duplicate handling
}
