% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_guess.R
\name{adam_guess}
\alias{adam_guess}
\title{Guess role columns for spec}
\usage{
adam_guess(role, type = c("bds", "occds"), colnames_data)
}
\arguments{
\item{role}{the spec role of interest}

\item{type}{either \code{bds} or \code{occds}. defaults to \code{bds}}

\item{colnames_data}{character vector, containing the column names of the
ADaM data set to be processed}
}
\value{
The guessed column name or NULL
}
\description{
Guess role columns for spec
}
