% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_adsl.R
\name{adam_spec_adsl}
\alias{adam_spec_adsl}
\title{Create spec object for adsl data sets}
\usage{
adam_spec_adsl(
  file,
  id = "SUBJID",
  trt = NULL,
  keep = NULL,
  drop = NULL,
  filter = NULL,
  attach_data = FALSE,
  ...
)
}
\arguments{
\item{file}{the path of the sas file to process}

\item{id}{name of id column to keep}

\item{trt}{name of trt column to keep (e.g. TRT01A)}

\item{keep}{columns to be kept (overrides blacklist)}

\item{drop}{superseded. columns to be dropped (overrides whitelist)}

\item{filter}{character vector of filter criteria to be evaluated}

\item{attach_data}{boolean. attach the imported raw data
adsl.sas7bdat' \%in\% list.files(path)}
}
\description{
Create spec object for adsl data sets
}
