% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_adsl.R
\name{adam_spec_adsl}
\alias{adam_spec_adsl}
\title{Create specification object for ADaM data sets of type 'adsl'}
\usage{
adam_spec_adsl(
  file = NULL,
  data = NULL,
  id = "SUBJID",
  trt = NULL,
  keep = NULL,
  drop = NULL,
  filter = NULL,
  fct_levels = NULL,
  attach_data = FALSE,
  catalog_file = NULL
)
}
\arguments{
\item{file}{the path of the sas(7bdat) or rds file to process}

\item{data}{tibble with the data in adsl format for which the specification
is created}

\item{id}{name of id (e.g. SUBJIDN, SUBJID) column to keep.
Highly redundant variables will not be included in the suggested set of columns returned in \code{select} (see Details).}

\item{trt}{column to be used as the treatment variable. All other predefined treatment variables (see Details) are added
to the \code{drop_list}. If NULL, all treatment variables will be added to the \code{drop_list}.}

\item{keep, drop}{columns to be kept/dropped, independent of the technical selection process within this function}

\item{filter}{character vector of filters following \code{dplyr::filter()} syntax for use in \code{\link{build_adsl}()} (see Details).
Defaults to NULL.}

\item{fct_levels}{optional list of named vectors providing code-decode pairs
and/or setting the level order (see details section for structure).}

\item{attach_data}{boolean. attach the imported raw data.}

\item{catalog_file}{path to the catalog file to be passed to
\code{\link[haven:read_sas]{haven::read_sas()}}. Defaults to NULL.
Ignored if \code{file} is not a sas7bdat file.}
}
\value{
A list containing the following
\item{\code{file}, \code{md5}}{the name and md5 checksum, resp., of the file the generated spec is based upon}
\item{\code{data}}{the raw data set if \code{attach_data}, NULL otherwise}
\item{\code{data_info}}{a list containing the number of subjects \code{nsubj} and columns \code{ncol} in the data after applying \code{filter}}
\item{\code{type}}{character string \code{adsl}, generally giving the type of ADaM data set processed (\code{adsl}/\code{bds}/\code{occds})}
\item{\code{filter}}{subset of \code{filter} that yields non-empty result when applied individually (using \code{\link{check_filter}()}}
\item{\code{select}}{the suggested list of columns to select from the data set}
\item{\code{factor_levels}}{a list containing a factor level code/decode for each column identified as a factor}
\item{\code{flag_table}}{a tibble with columns id and any columns identified as flag (character and matching numeric) based on matching column names or labels}
\item{\code{id}, \code{trt}}{passing unchanged input}
\item{\code{drop_list}}{a list containing column names suggested to be dropped with the entry
name identifying the rationale for the discard
\describe{
\item{\code{drop}}{passing the user input \code{drop}}
\item{\code{datetime}}{date/times columns}
\item{\code{numcode}}{numeric code for another variable (incl numeric flags)}
\item{\code{flag}}{flags (both numeric and character columns), see also \code{flag_table}}
\item{\code{combination}, \code{empty}, \code{constant}}{combined, empty and constant columns, resp.}
\item{\code{redundancy}}{columns with redundant information to \code{id} and \code{trt} if provided)}
}}
\item{\code{spec_id}}{character string \code{adsl}, generally the name of the domain}
\item{\code{dict}}{a tibble of column names and labels (if present in the data set)}
}
\description{
Given a file containing an adsl data set, \code{\link{adam_spec_adsl}()} will create a specification
object for use in \code{\link{build_adsl}()} to actually create a subset of
the data to be used in machine learning. For adsl specifically, the main task is the
identification of noise and redundancies in the data and the selection of a potentially meaningful
set of columns (returned in \code{select}) and redundancies in the data.
}
\details{
\describe{
\item{\emph{Subject id}}{Non-numeric columns are recoded as numeric, based on the order in which they appear in the data
(sorted by \code{id}). All columns with a perfect Spearman correlation to \code{id} are considered redundant and added to
the \code{drop_list}. In addition, all numeric columns with a perfect Spearman correlation to RANDDT (if available in the data)
are also added to the \code{drop_list}, as well as RANDNO (if present in data).}
\item{\emph{Treatment variable}}{The predefined list of treatment variables is TRT01A, ARMCD, ARM, ACTARM, ACTARMCD, TRT01P, TR01PG1, TR02PG1, TR01AG1, TR02AG1.
No more than one of these variables will be returned in \code{select}. Note that the chosen treatment representing
variable will be renamed to the standard '.trt' in \code{\link{build_adsl}()}.}
\item{\emph{Filter check}}{Filters will be checked against the data and will only be kept if the filter would not throw an error and if the resulting
data set has positive number of rows. See \code{\link{check_filter}()} for further details.}
\item{\emph{fct_levels}}{\code{adam_spec_adsl()} will try and derive the factor levels from
the data set by identifying column of code/decode pairs using a simple
heuristic and any formats present in the optionally provided catalog file.
For ADaM 2.0, the number of the code/decode column pairs is expected to
be reduced to a minimum and mainly numeric codes are expected to be present.
\code{fct_levels} can be used to ensure that the columns are treated as factors
in the first place and to provide the factor labels manually for
interpretability of the results.
\code{fct_levels} is provided as a named list, containing one entry per factor
column, that should be defined/updated. Each entry is a named vector with
the names being the level names and values the corresponding
entries used in the actual data.}
}
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

