% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_spec.R
\name{adjust_spec}
\alias{adjust_spec}
\title{Adjust spec object}
\usage{
adjust_spec(spec, id, ..., append = FALSE)
}
\arguments{
\item{spec}{spec object to modify}

\item{id}{name of list element to modify in the spec}

\item{...}{modifications to the \code{spec[[id]]} of the form \verb{<name> = <value>}.}

\item{append}{boolean. If TRUE, modifications are appended, otherwise overwritten. defaults to FALSE.}
}
\value{
A modified version of \code{spec} to be used as input to \verb{\link{build}()}
}
\description{
Helper function to make adjustments to the spec object built by \verb{\link{adam_spec}()} to be used with the \verb{\%>\%}.
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

