% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{build}
\alias{build}
\title{Build the feature matrix from various sources according to a specification object}
\usage{
build(spec, filter = NULL, keep = NULL, drop = NULL, join = dplyr::inner_join)
}
\arguments{
\item{spec}{a specification object as provided by \code{\link{adam_spec}()} (either \code{spec} or \code{path} has to be provided)}

\item{filter}{a character vector of conditions to be passed to \code{dplyr::filter()},
e.g. regarding visits, treatment arms or parameters. Defaults to NULL. Only applied, if \code{spec} is not provided.}

\item{keep}{character vector defining the subset of data sets in the given \code{path} to create
the specification for (e.g. \code{c('adsl', 'advs'))}). If both \code{keep} and \code{drop} are specified,
\code{keep} overrides \code{drop}. Defaults to NULL. Only applied, if \code{spec} is not provided.}

\item{drop}{character vector defining a subset of data sets in the given \code{path} to
be excluded from the list of specifications (e.g. \code{'adqseq5d')}). Defaults to NULL.
Only applied, if \code{spec} is not provided.}

\item{join}{either function to join data sets (e.g. \code{dplyr::full_join()} or a character (vector) giving the names
of the data sets containing the .ids to keep (e.g. \code{join = c('adxb', 'adlb')}). defaults to \code{dplyr::inner_join}}
}
\value{
\code{build()} returns a wide data set with one row per subject and standardized column names for the subject id (\code{.id})
and the treatment variable (\code{.trt}), if it is provided in the \code{spec} object. Objects with additional information on
the data are provided in the attributes of the returned object.

\item{\code{dict}}{
\describe{
\item{\code{param}}{original parameter name in the source data}
\item{\code{column}}{column name of the variable in the returned data. \code{column} is derived from \code{param} by transforming
it into a valid file name and possibly adding a time extension, if multiple time points are considered for a particular parameter.}
\item{\code{label}}{parameter label}
\item{\code{source}}{source id provided by the specification object. If created with \code{\link{adam_spec}()}, this is the name of the domain.}
\item{\code{type}}{adam data type of the source data (adsl, bds or occds)}
\item{\code{unit}}{parameter unit (if applicable)}
\item{\code{time}}{measurement time point (if applicable)}
}}
\item{\code{source}}{file path and md5 checksums of the source data sets}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#maturing}{\figure{lifecycle-maturing.svg}{options: alt='[Maturing]'}}}{\strong{[Maturing]}}

The \code{build()} function allows to build a machine learning data set from a specification object as provided
by \code{\link{adam_spec}()} (with or without data already attached).
}
\details{
Missing values in variables from occurrence data sets are interpreted as 'absence of event',
whereas NAs in adsl and bds data are considered to be true missing values.
For missing values in occds data after joining with other data sets,
missing values are replace by 0 for numerics, an additional level 'none' is introduced for
for factors.
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

\seealso{
\code{\link{build_adsl}()}, \code{\link{build_bds}()}, \code{\link{build_occds}()}
}
