% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_adsl.R, R/build_bds.R, R/build_occds.R,
%   R/build_x.R
\name{build_adsl}
\alias{build_adsl}
\alias{build_bds}
\alias{build_occds}
\alias{build_x}
\title{Create wide format data following a given spec}
\usage{
build_adsl(spec)

build_bds(
  spec,
  dupl_ctrl = list(values_fn = NULL, arrange = NULL),
  names_ctrl = list(clean_fn = ~stringr::str_replace_all(.x, "[:punct:]|[:space:]", "_"),
    names_sep = "_"),
  rm = FALSE
)

build_occds(spec, values_fn_occds = NULL)
}
\arguments{
\item{spec}{result of \verb{adam_spec_*()}}

\item{dupl_ctrl}{bds only. A list with two entries
\itemize{
\item{\code{values_fn}} function to handle duplicates in pivoting step. see details section for default.
\item{\code{arrange}} expression passed to \code{arrange()} optional sorting of data set prior to pivoting,
e.g. in order to select the first/last value by date. defaults to \code{NULL}.
}}

\item{names_ctrl}{bds only. A list with two entries handling cleaning and renaming of columns after pivoting
\itemize{
\item{\code{clean_fn}} defaults to \code{stringr::str_replace_all(.x, '[:punct:]|[:space:]', '_')}.
\item{\code{names_sep}} defaults to '_'
}}

\item{rm}{bds only. boolean. defaults to FALSE. if TRUE, a repeated measurement feature matrix with an
additional \code{.rmtime} column is prepared. Only used, if \code{is.null(spec$rm)}.}

\item{values_fn_occds}{occds only. function that is used to summarize values in the pivoting step,
if multiple rows per observation unit are present. If \code{NULL}, the maximum is used for numeric values and
the last factor level is used for categorical values.}
}
\value{
A list with the following entries
\itemize{
\item \code{data} a tibble in wide format with one row per \code{id}
\item \code{dict} a tibble listing the distinct combinations of columns
\code{param}, \code{label}, \code{unit}, \code{time}, \code{column}, \code{source} (if provided).
\item \code{source} a list passing the \code{file} slot from the given \code{spec} that the created data set is based
upon along with the md5 checksum of this file if \code{file} was provided, NULL otherwise
\item \code{flag_table} \code{build_adsl()} only. flag table is passed from \code{spec$flag_table} slot
}
}
\description{
Extract and reshape data from a single data set according to the given specification
as created by \verb{adam_spec_*()}. In addition, an (updated) dictionary is returned along with
the md5 checksum of the specified file.
}
\details{
Note that the output dictionary may differ from the dictionary created by \verb{adam_spec_*()},
as multiple features may be derived from a single parameter at different time points.

\code{values_fn} is passed to \code{pivot_wider()}.
The default is \code{function(x) {ifelse(all(is.numeric(x)), mean(x, na.rm = TRUE), na.omit(x)[1])}}
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

