% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_adsl.R, R/build_bds.R, R/build_occds.R,
%   R/build_x.R
\name{build_adsl}
\alias{build_adsl}
\alias{build_bds}
\alias{build_occds}
\alias{build_x}
\title{Create wide format data following a given spec}
\usage{
build_adsl(spec)

build_bds(
  spec,
  values_fn = function(x) {     ifelse(all(is.numeric(x)), mean(x, na.rm = TRUE), x[1])
    }
)

build_occds(spec)
}
\arguments{
\item{spec}{result of \verb{adam_spec_*()}}

\item{values_fn}{bds only. function to handle duplicates in pivoting step. see details section for default.}
}
\value{
A list with the following entries
\item{\code{data}}{a tibble in wide format which one row per \code{id}}
\item{\code{dict}}{a tibble listing the distinct combinations of columns
\code{param}, \code{label}, \code{unit}, \code{time}, \code{column}, \code{source} (if provided).}
\item{\code{source}}{a list passing the \code{file} slot from the given \code{spec} that the created data set is based
upon along with the md5 checksum of this file}
\item{\code{flag_table}}{\code{build_adsl()} only. flag table is passed from \code{spec$flag_table} slot}
}
\description{
Extract and reshape data from a single data set according to the given specification
as created by \verb{adam_spec_*()}. In addition, an (updated) dictionary is returned along with
the md5 checksum of the specified file.
}
\details{
Note that the output dictionary may differ from the dictionary created by \verb{adam_spec_*()},
as multiple features may be derived from a single parameter at different time points.

\code{values_fn} is passed to \code{pivot_wider()}. The default is \code{function(x) {ifelse(all(is.numeric(x)), mean(x, na.rm = TRUE), x[1])}}
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

