% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_filter.R
\name{check_filter}
\alias{check_filter}
\title{Check filter}
\usage{
check_filter(data, filter, data_id = NULL)
}
\arguments{
\item{data}{data set that the given filter(s) should be applied to}

\item{filter}{an expression to be used in \code{dplyr::filter()}, (either a single filter, or multiple ones separated by ',')}

\item{data_id}{character to include in warning message to help identify the data sets when used in \code{\link{adam_spec}()}. defaults to NULL.}
}
\value{
A list with two entries: \code{individual} for single filter assessment,
\code{overall_norow} for the combined filter assessment, if all individually applicable filters are applied to \code{data}.
\code{Individual} consists of a list of three logicals per filter, with values TRUE
if the application of the \emph{individual} filter to \code{data} yields
\describe{
\item{is_error}{an error}
\item{is_norow}{a tibble with 0 rows}
\item{keep}{neither of the above}
}
\code{overall_norow} is TRUE, if the combination of all applicable filters to \code{data} results in a 0-row tibble. In this case a warning is thrown.
}
\description{
Check a given (set of) filter(s) against a data set and assess whether or not a non-empty data set would be returned.
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

