% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_freq.R
\name{check_freq}
\alias{check_freq}
\title{Identify factors with low frequency classes}
\usage{
check_freq(x, thres = 10)
}
\arguments{
\item{x}{output of \verb{\link{prepare_ml}()} or a tibble}

\item{thres}{integer. Factors with at least one class of size smaller than \code{thres} will be identified. Defaults to 10.}
}
\value{
In case factors are present in the data set that meet the condition of interest,
relevant names are printed in the console and a list of frequency tables is returned.
}
\description{
Identify factors with low frequency classes
}
\details{
If \code{x} is the output of \verb{\link{prepare_ml}()}, the tibble \code{x$data$prep$train} is checked for
factors with at least one low frequency class.
Please refer to the package vignette for further details.
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

