% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_helpers.R
\name{check_occds_occur}
\alias{check_occds_occur}
\title{Check for 'N' values in --OCCUR column in occds}
\usage{
check_occds_occur(
  data,
  domain = NULL,
  filters = NULL,
  quiet = FALSE,
  no_char = "N",
  no_num = 0
)
}
\arguments{
\item{data}{data set to check}

\item{domain}{name of data set}

\item{filters}{filters to be applied before checking for issues.
defaults to NULL, in which case no filters are applied.}

\item{quiet}{whether to suppress messaging in the console. defaults to FALSE.}

\item{no_char, no_num}{values that code for 'no' in pre-specified list of
events. defaults to \code{N} and \code{0} for character and numeric variables, resp.}
}
\value{
invisibly returns character vector of potentially problematic columns.
(Invisible) return value has length 0 if no problems were detected.
}
\description{
Check for 'N' values in --OCCUR column in occds
}
