% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_helpers.R
\name{create_dict}
\alias{create_dict}
\title{Create dictionary for spec entry}
\usage{
create_dict(spec_entry)
}
\arguments{
\item{spec_entry}{Top level entry of an object of class \code{martini_spec}.}
}
\value{
the dictionary as a tibble.
For all data types, the dict contains columns
\code{label}, \code{source}, \code{type} and \code{selected}.
In addition for data type adsl and bds there is \code{param}, with
\code{unit} and \code{selected} as additional column for bds and adsl, resp.
}
\description{
Create dictionary for spec entry
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

