% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_helpers.R
\name{data_info}
\alias{data_info}
\title{data info on spec id}
\usage{
data_info(spec_entry)
}
\arguments{
\item{spec_entry}{Top level entry of an object of class \code{martini_spec}.}
}
\value{

}
\description{
info on size of extracted data in terms of numbers of subjects and columns (roughly)
}
\examples{
data_info(martini_spec$adsl)
}
