% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adam_spec_helpers.R
\name{import_info}
\alias{import_info}
\title{Import file and collect info}
\usage{
import_info(file, catalog_file = NULL)
}
\arguments{
\item{file}{filepath}

\item{catalog_file}{path to the catalog file to be passed to
\code{\link[haven:read_sas]{haven::read_sas()}}. Defaults to NULL.
Ignored if \code{file} is not a sas7bdat file.}
}
\value{
list containing data and corresponding md5 sum and file size
}
\description{
Import file and collect info
}
