% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bds.R
\name{pivot_prepare_bds}
\alias{pivot_prepare_bds}
\title{Prepare bds data for pivoting step in build}
\usage{
pivot_prepare_bds(
  bds_full,
  spec,
  values_fn = NULL,
  arrange = NULL,
  clean_fn = ~stringr::str_replace_all(.x, "[:punct:]|[:space:]", "_"),
  names_sep = "_",
  rm = FALSE
)
}
\arguments{
\item{bds_full}{original bds-type data set}

\item{spec}{Top level entry of an object of class \code{martini_spec}.}

\item{values_fn}{function to control duplicate handling in \code{pivot wider()}.
See docu of \code{build_bds()} for details.}

\item{arrange}{expression to be passed to \code{dplyr::arrange()} before
pivoting. Relevant in case of duplicates and if \code{values_fn} is
sensitive to the order of values to aggregate.}

\item{clean_fn}{function to clean future column names (after pivoting), defaults to
\code{~ stringr::str_replace_all(.x, '[:punct:]|[:space:]', '_')}}

\item{names_sep}{to be passed to \code{pivot wider()}. defaults to \verb{_}.}

\item{rm}{boolean. defaults to FALSE. if TRUE, pivoting for a repeated measurement feature matrix with an
additional \code{.rmtime} column is prepared. Only used, if \code{is.null(spec$rm)}.}
}
\value{
A list containing the pivot_wider arguments (pivot_args) as well as the function
to clean column names (clean_fn). The \code{pivot_args} list includes the prepared data set (filtered, arranged)
as well as pivot_wider params (key(s), value, values_fn, names_sep)
}
\description{
Preparation of dataset bds_full as well as parameters to be passed to pivot_wider
in build_bds to allow for appropriate unit testing
}
\details{
Data preparation of bds_full for pivoting includes filtering and arranging
the data set before relevant columns are selected and renamed using \code{clean_fn} (\code{param}, \code{time} only)
If the prepared data set has more than one level in the \code{time} column,
names_from will be a vector of the form \code{c(param, time)}
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

