% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ml_vars.R
\name{prepare_ml_feature}
\alias{prepare_ml_feature}
\title{prep feature matrix}
\usage{
prepare_ml_feature(feature, vars_fct_expl_na = NULL, level_other = "other")
}
\arguments{
\item{feature}{feature tibble}

\item{vars_fct_expl_na}{defaults to NULL}

\item{level_other}{defaults to 'other'}
}
\value{
updated feature matrix
}
\description{
prep feature matrix
}
\details{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated since update in \code{prepare_ml()} as of martini 0.6.0
}
