% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ml_vars.R
\name{prepare_ml_feature}
\alias{prepare_ml_feature}
\title{prep feature matrix}
\usage{
prepare_ml_feature(feature, vars_fct_expl_na = NULL, level_other = "other")
}
\arguments{
\item{feature}{}
}
\value{
updated feature matrix
}
\description{
prep feature matrix
}
