% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ml_outcome.R
\name{prepare_ml_outcome}
\alias{prepare_ml_outcome}
\title{Prepare ML ready outcome data set}
\usage{
prepare_ml_outcome(
  outcome,
  outcome_name = NULL,
  level_order = NULL,
  outlier_remove = FALSE,
  outlier_ctrl = list(coef = 3)
)
}
\arguments{
\item{outcome}{tibble containing \code{.id} column and the outcome of interest}

\item{outcome_name}{if \code{NULL} (default), the first column that's not \code{.id} is chosen for \code{outcome_name}
and the \code{outcome_type} is guessed to be either classification or regression.
One may also provide a single character giving the name of the outcome column OR a vector of length two giving the
column names for the 'time' and 'status' data in survival analysis, where \code{.time} is numeric and \code{.status}
is binary with 0 coding for censored, and 1 coding for event. Currently, only right-censoring is supported. Please note,
that survival will never be guessed.}

\item{level_order}{Level order for a classification outcome. \code{NULL} keeps the natural order (only used for classification).}

\item{outlier_remove}{Remove outliers in a regression outcome based on the 'boxplot definition'.
The outlier coefficient can be modified in \code{outlier_ctrl}, defaults to 3 (only used for regression).}

\item{outlier_ctrl}{Control list for the outlier removal, if \code{outlier_remove} is \code{TRUE}. Currently, the list contains only
the boxplot outlier coefficient \code{coef}, which defaults to 3.}
}
\value{
A list with the following slots

\item{\code{outcome}}{The outcome data set containing only the id and one or two columns
with standardized column names (\code{.out} for regression or classification, \code{.time} and \code{.status} for survival).}
\item{\code{outcome_name}}{Named vector with the original name(s) of the outcome variable(s).}
\item{\code{outcome_label}}{Named vector with the labels(s) of the outcome variable(s). If the columns of \code{outcome} do not contain labels,
the column name is used instead.}
\item{\code{outcome_mode}}{The outcome mode (\code{regression}, \code{classification} or \code{survival}.}
\item{\code{outcome_dict}}{Dictionary tibble for the outcome variable(s).}
\item{\code{na_outcome}}{The IDs of NAs in \code{outcome}.}
\item{\code{id_outlier}}{The IDs of removed outliers.}
}
\description{
Prepares an ML ready outcome data set (used in \code{\link{prepare_ml}()})
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

