% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ml_recipe.R
\name{prepare_ml_recipe}
\alias{prepare_ml_recipe}
\title{create recipe from data}
\usage{
prepare_ml_recipe(
  data,
  prep_recipe = NULL,
  corr_method = "pearson",
  corr_use = "pairwise.complete.obs",
  thres_list = NULL,
  step_list = NULL,
  vars_imp_ignore = c(".trt"),
  vars_fct_expl_na = NULL,
  vars_ordinalscore = NULL,
  vars_keep_corr = NULL,
  vars_no_trafo = NULL,
  one_hot,
  log_base
)
}
\arguments{
\item{data}{raw data set to create recipe for}

\item{prep_recipe}{if NULL, recipe will be created}

\item{corr_method, corr_use}{defaulting to corr_method "pearson" and corr_use "pairwise.complete.obs"}

\item{thres_list, step_list}{named list objects collecting all threshold values
and step selection info, resp. please refer to the documentation of the \verb{thres_*} and \verb{prep_step_*} arguments
in \code{\link{prepare_ml}()} for detailed documentation and list entry names.}

\item{vars_imp_ignore}{variables that shall not be imputed can be specified
in \code{vars_imp_ignore} (vector of column names, defaults to
\code{vars_imp_ignore = '.trt'}). Observations with missing values in these
variables will be removed. Removal is documented in \code{removed$rows}.}

\item{vars_fct_expl_na}{column names of factors for which NAs should be
treated as an explicit factor level. Defaults to NULL.}

\item{vars_ordinalscore}{column names of ordinal factor variables to be
converted into numeric scores. Defaults to NULL.}

\item{vars_keep_corr}{choose these variables over other options when removing
variables due to high correlation in \code{recipes::step_corr()}.
See \code{recipes::step_rm()} below for details.}

\item{vars_no_trafo}{character vector defining variables that should be
excluded from transformation steps such as log transformation and/or normalization
(if applicable). Defaults to NULL.}

\item{one_hot}{boolean. passed to \code{recipes::step_dummy()} to choose one
hot encoding over dummy encoding}

\item{log_base}{base to use for log-transformation in
\code{recipes::step_log()}. Defaults to \emph{exp(1)}.}
}
\value{
a named list with entries containing
\itemize{
\item the prepared recipe
\item info on steps included in the recipe
\item a list of relevant variables
\item a list of thresholds used
\item \code{high_corr} a tibble listing correlations above \code{thres_corr}. \code{NULL} if \code{step_list$prep_step_corr = FALSE}.
}
}
\description{
create recipe from data
}
\seealso{
\code{\link{prepare_ml}()}
}
