% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_ml_vars.R
\name{prepare_ml_vars}
\alias{prepare_ml_vars}
\title{Prepare ML helper function}
\usage{
prepare_ml_vars(
  data,
  thres_count = NULL,
  thres_log = NULL,
  thres_lump = NULL,
  thres_imp = NULL,
  remove = c(".id", ".out", ".status", ".time", ".trt")
)
}
\arguments{
\item{data}{the data set to be searched for feature sets with specific characteristics relevant for further data preparation}

\item{thres_count}{used to detect integer columns with up to \code{thres_count} distinct values (might be excluded from further processing, e.g. log & normalization)}

\item{thres_log}{threshold for log transformation}

\item{thres_lump}{proportion threshold for factor lumping; used to detect factors with exactly one level having a relative frequency below \code{thres_lump}}

\item{thres_imp}{proportion threshold used to detect variables with non-missing proportion up to \code{thres_imp}}

\item{remove}{columns to be excluded from all identified sets; defaults to c(".id", ".out", ".status", ".time")}
}
\value{
A list with slots specifying the detected variable sets of interest.
NA if required thresholds were not defined; NULL if no variables meet the corresponding criteria.

\item{count}{assumed to be counts}
\item{log}{to be log transformed as the skewness exceeds \code{thres_log}}
\item{nolump}{to be excluded from lumping}
\item{imp}{to be imputed}
\item{exclude}{to be excluded from the data as the proportion of missing values exceeds \code{thres_imp}}
}
\description{
Identify variable sets from input matrix that might require extra steps in data preparation,
e.g. skewed variables to be log transformed, counts
}
\section{Authors}{

Maike Ahrens (ahrensmaike), Sebastian Voss (svoss09)
}

