% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_corr_keep.R
\name{step_corr_keep}
\alias{step_corr_keep}
\title{High correlation filter with a twist}
\usage{
step_corr_keep(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  threshold = 0.9,
  use = "pairwise.complete.obs",
  method = "pearson",
  keep = NULL,
  removals = NULL,
  high_corr = NULL,
  columns_logged = NULL,
  skip = FALSE,
  id = recipes::rand_id("corr_keep")
)
}
\arguments{
\item{threshold}{A value for the threshold of absolute correlation values.
The step will try to remove the minimum number of columns so that all the
resulting absolute correlations are less than this value.}

\item{use}{A character string for the \code{use} argument to the \code{\link[stats:cor]{stats::cor()}}
function.}

\item{method}{A character string for the \code{method} argument to the
\code{\link[stats:cor]{stats::cor()}} function.}

\item{removals}{A character string that contains the names of columns that
should be removed. These values are not determined until \code{\link[=prep]{prep()}} is
called.}
}
\description{
\code{step_corr_keep()} creates a \emph{specification} of a recipe step that will
potentially remove variables that have large absolute correlations with other
variables BUT with a twist.
}
\details{
This step attempts to remove variables to keep the largest absolute
correlation between the variables less than \code{threshold}.

The filter tries to prioritize predictors for removal based on the global
affect on the overall correlation structure. If you have two identical
predictors, the variable ordered first will be removed.

When a column has a single unique value, that column will be excluded from
the correlation analysis. Also, if the data set has sporadic missing values
(and an inappropriate value of \code{use} is chosen), some columns will also be
excluded from the filter.

The arguments \code{use} and \code{method} don't take effect if case weights are used
in the recipe.
}
\section{Tidying}{
\describe{
\item{terms}{character, the selectors or variables selected to be removed}
\item{id}{character, id of this step}
}

\if{html}{\out{<div class="sourceCode {r, echo = FALSE, results="asis"}">}}\preformatted{step <- "step_corr"
result <- knitr::knit_child("man/rmd/tunable-args.Rmd")
cat(result)
}\if{html}{\out{</div>}}
}

\examples{
\dontshow{if (rlang::is_installed("modeldata")) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(biomass, package = "modeldata")

set.seed(3535)
biomass$duplicate <- biomass$carbon + rnorm(nrow(biomass))

biomass_tr <- biomass[biomass$dataset == "Training", ]
biomass_te <- biomass[biomass$dataset == "Testing", ]

rec <- recipe(
  HHV ~ carbon + hydrogen + oxygen + nitrogen + sulfur + duplicate,
  data = biomass_tr
)

corr_filter <- rec |>
  step_corr_keep(all_numeric_predictors(), threshold = .5)

filter_obj <- prep(corr_filter, training = biomass_tr)

filtered_te <- bake(filter_obj, biomass_te)
round(abs(cor(biomass_tr[, c(3:7, 9)])), 2)
round(abs(cor(filtered_te)), 2)

tidy(corr_filter, number = 1)
tidy(filter_obj, number = 1)
\dontshow{\}) # examplesIf}
}
\author{
Modified from recipes. Original R code for filtering algorithm by Dong Li, modified by Max
Kuhn. Contributions by Reynald Lescarbeau (for original in \code{caret}
package). Max Kuhn for the \code{step} function.
}
